API disruption as a service.
The registry API can now handle routing to api.npmjs.org or to the
registry of your choice. It's no longer a singleton but an object
you construct & hand the yargs argv to. That should configure it
properly 90% of the time.

Updated the `whoami` command to use the new API. Updated tests.
Documented the api in the registry.js file.